Refactor code by adding utils folder, extract code block into create.rb file and modify app.rb, main.rb file for the app to work properly
