+++ utils/list.rb
@@ -1,0 +1,4 @@
+require './classes/student'
+require './classes/teacher'
+require './classes/book'
+require './classes/rental'
